Refactored nodetype XML handling (preparing for JCR-54)
The load and store methods in NodeTypeDefStore are now implemented
using separate formatter classes in the new o.a.j.core.nodetype.xml
package. The changes are highly localized, pass all unit tests, and
have little externally visible effects.

The externally visible changes are:

   * Much more strict XML validation (missing or mistyped attributes
     now cause exceptions instead of backing up to default values)
   * Exception messages and root causes may be different
   * The ChildItemDef class is public

The changes have been approved by Stefan.

In addition to cleaning and documenting existing functionality, this
change also prepares for the work of removing the JDOM dependency as
discussed in JCR-54.

git-svn-id: https://svn.apache.org/repos/asf/incubator/jackrabbit/trunk@157342 13f79535-47bb-0310-9956-ffa450edef68