JCR-54: Removed JDOM dependency from PredefinedNodeTypeTest.
In addition to removing the JDOM dependency, this commit changes the underlying design
of this test case. Instead of parsing a node type definition file and comparing the parsed
results to the existing node types, the test case now generates a "spec string" from the
node type and compares it with the node type definition extracted from the JSR 170 specification.

git-svn-id: https://svn.apache.org/repos/asf/incubator/jackrabbit/trunk@159677 13f79535-47bb-0310-9956-ffa450edef68