refactor the version of createResource() which calls resource.setIsCollection() so that the decision when to call that method and the calculation of the value passed to it are delegated to helper methods.
isCreateRequest() returns true if the request method signifies that a
resource needs to be created. this is the circumstance in which
resource.setIsCollection() needs to be called. the default
implementation of isCreateRequest() returns true when the method is
MKCOL, PUT or POST.

isCreateCollectionRequest() erturns true if the request method
signifies that a *collection* resource needs to be created. it is used
to decide whether to pass true or false to
resource.setIsCollection(). the default implementation returns true
when the method is MKCOL.

this refactoring allows subclasses to override/extend the logic used
for making these choices. for example, cosmo's
CosmoResourceFactoryImpl extends the methods to return true when the
http request method is MKCALENDAR.

we also add accessors for the lockManager and resourceConfig
attributes and make them non-final so that other objects can modify
them.


git-svn-id: https://svn.apache.org/repos/asf/incubator/jackrabbit/trunk/contrib/jcr-server@354822 13f79535-47bb-0310-9956-ffa450edef68