JCR-1171: Contribute Pluggable Permission and User Management to Jackrabbit
Moved the o.a.j.commons.iterator.AccessControlPolicyIteratorAdapter class to o.a.j.core.security as it is only accessed from there and we currently (before JSR 283 is final) can't put the class in jcr-commons.

With this change all jackrabbit-core classes are included in o.a.j.core, which makes things easier to track.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@705934 13f79535-47bb-0310-9956-ffa450edef68