JCR-1952: DOMException: NAMESPACE_ERR thrown when exporting document view
Workaround to the XmlReader failure in the TCK. It seems like the TCK can only load an XmlReader if a TransformerFactory has been loaded before. Accessing the SerializingContentHandler class is one way to force this, as the static initializer of that class uses a TransformerFactory.

TODO: Should fix the TCK...

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@746666 13f79535-47bb-0310-9956-ffa450edef68