JCR-2120: java.lang.Iterable support for RangeIterators
Use a NodeIterable adapter on the result of the NodeIterator return value. This way potential RepositoryExceptions are thrown directly and don't need to be encapsulated in RuntimeExceptions. And the code is cleaner!

The only downside is that the returned Iterable always returns the same Iterator from the iterator() method, but that "feature" has been documented in the NodeIterable javadoc.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@776681 13f79535-47bb-0310-9956-ffa450edef68