JCR-2762: Optimize bundle serialization
Refactor serialization and deserialization code from ItemStateBinding and BundleBinding to separate BundleReader and BundleWriter classes. This makes it easier to later on track the bundle version as it can be kept as an instance variable of an BundeReader object.

Make the use of DataInputStream in readBundle() and checkBundle() an implementation detail that's not exposed in the method signature.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@1002707 13f79535-47bb-0310-9956-ffa450edef68