JCR-2699: Improve read/write concurrency
Even the tiny synchronized block in the LRU cache becomes a source of lock contention, so replace it with a segmented cache that has now single synchronization block over the entire cache. The downside is a slight deviation from the LRU eviction policy.

Replaced the BundleCache and LRUNodeIdCache classes with the new ConcurrentCache implementation. Instead of using a separate data structure, a special MISSING sentinel bundle is used to to mark non-existent bundles.

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@1004182 13f79535-47bb-0310-9956-ffa450edef68